Fix total hours count #5
Addresses issue #5: Get total_hours returns empty data set
- Query type-casts total_hours as a float
- Model has been changed to float64

Resolves #5

+++ api/models.go
@@ -25,8 +25,8 @@ type DBWaterOutage struct {
 	OutageType string `json:"outage_type,omitempty"`
 	CreatedAt string `json:"created_at,omitempty"`
 	UpdatedAt string `json:"updated_at,omitempty"`
-	TotalOutages uint8 `json:"total_outages,omitempty"`
-	TotalHours int `json:"total_hours,omitempty"`
+	TotalOutages int `json:"total_outages,omitempty"`
+	TotalHours float64 `json:"total_hours,omitempty"`
 }
 
 // DBWaterOutageCol returns a reference for a column of a DBWaterOutage
